# pkg/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the ISDuBA API.",
        "title": "ISDuBA API",
        "contact": {
            "name": "ISDuBA team",
            "url": "https://github.com/ISDuBA/ISDuBA",
            "email": "info@intevation.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Returns all events that match the specified query.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns a list of events.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event query",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.overviewEvents.events"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/events/{publisher}/{trackingid}": {
            "get": {
                "description": "Returns all events from the specified advisory.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns all events.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Publisher",
                        "name": "publisher",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tracking ID",
                        "name": "trackingid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/web.viewEvents.event"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Event": {
            "type": "string",
            "enum": [
                "import_document",
                "delete_document",
                "state_change",
                "add_sscv",
                "change_sscv",
                "delete_sscv",
                "add_comment",
                "change_comment",
                "delete_comment"
            ],
            "x-enum-comments": {
                "AddCommentEvent": "AddCommentEvent represents the addition of a comment.",
                "AddSSVCEvent": "AddSSVCEvent represents the addtion of a SSVC score.",
                "ChangeCommentEvent": "ChangeCommentEvent represents the change of a comment.",
                "ChangeSSVCEvent": "ChangeSSVCEvent represents the change of a SSVC score.",
                "DeleteCommentEvent": "DeleteCommentEvent represents the deletion of a comment.",
                "DeleteDocumentEvent": "DeleteDocumentEvent represents a document deletion.",
                "DeleteSSVCEvent": "DeleteSSVCEvent represents the deletion of a SSVC score.",
                "ImportDocumentEvent": "ImportDocumentEvent represents a document import.",
                "StateChangeEvent": "StateChangeEvent represents changing the advisory state."
            },
            "x-enum-varnames": [
                "ImportDocumentEvent",
                "DeleteDocumentEvent",
                "StateChangeEvent",
                "AddSSVCEvent",
                "ChangeSSVCEvent",
                "DeleteSSVCEvent",
                "AddCommentEvent",
                "ChangeCommentEvent",
                "DeleteCommentEvent"
            ]
        },
...
        "models.Workflow": {
            "type": "string",
            "enum": [
                "new",
                "read",
                "assessing",
                "review",
                "archived",
                "delete"
            ],
            "x-enum-comments": {
                "ArchivedWorkflow": "ArchivedWorkflow represents 'archived'.",
                "AssessingWorkflow": "AssessingWorkflow represents 'assessing',",
                "DeleteWorkflow": "DeleteWorkflow represents 'delete'.",
                "NewWorkflow": "NewWorkflow represents 'new'.",
                "ReadWorkflow": "ReadWorkflow represents 'read'.",
                "ReviewWorkflow": "ReviewWorkflow represents 'review'."
            },
            "x-enum-varnames": [
                "NewWorkflow",
                "ReadWorkflow",
                "AssessingWorkflow",
                "ReviewWorkflow",
                "ArchivedWorkflow",
                "DeleteWorkflow"
            ]
        },
...
        "web.overviewEvents.events": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        },
...
        "web.viewEvents.event": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string"
                },
                "comment_id": {
                    "type": "integer"
                },
                "document_id": {
                    "type": "integer"
                },
                "event_type": {
                    "$ref": "#/definitions/models.Event"
                },
                "state": {
                    "$ref": "#/definitions/models.Workflow"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
    }
}